Move metrics to a separate package
changelog_begin
changelog_end
