+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/metrics/MetricsSet.scala
@@ -15,6 +15,7 @@ import com.google.protobuf.timestamp.Timestamp
 
 import java.time.{Clock, Duration}
 import scala.concurrent.duration.FiniteDuration
+import com.daml.ledger.api.benchtool.metrics.metrics._
 
 object MetricsSet {
