+++ ledger/ledger-api-bench-tool/src/main/scala/com/daml/ledger/api/benchtool/metrics/metrics/TotalCountMetric.scala
@@ -1,9 +1,11 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.benchtool.metrics
+package com.daml.ledger.api.benchtool.metrics.metrics
 
 import java.time.Duration
+
+import com.daml.ledger.api.benchtool.metrics.{Metric, MetricValue}
 
 final case class TotalCountMetric[T](
     countingFunction: T => Int,
